RDoc-2532 Client API > Operations > Maintenance > Identities > Get identities (Part 2)
