+++ Documentation/5.2/Raven.Documentation.Pages/client-api/operations/maintenance/identities/get-identities.java.markdown
@@ -1,4 +1,4 @@
-# Get Identities Operations
+# Get Identities Operation
 
 **GetIdentitiesOperation** is used to return a dictionary which maps from the collection name to the identity value.
